Forbid import loops with eslint import/no-cycle rule
